+++ src/contour/ContourGuiApp.cpp
@@ -277,13 +277,13 @@ int ContourGuiApp::terminalGuiAction()
     QCoreApplication::setOrganizationName("contour");
     QCoreApplication::setApplicationVersion(CONTOUR_VERSION_STRING);
 
-// NB: High DPI scaling should be enabled, but that sadly also applies to QOpenGLWidget
-// which makes the text look pixelated on HighDPI screens. We want to apply HighDPI
-// manually in QOpenGLWidget.
-// QCoreApplication::setAttribute(Qt::AA_EnableHighDpiScaling, true);
-#if QT_VERSION < QT_VERSION_CHECK(6, 0, 0)
-    QCoreApplication::setAttribute(Qt::AA_DisableHighDpiScaling);
-#endif
+    // NB: High DPI scaling should be enabled, but that sadly also applies to QOpenGLWidget
+    // which makes the text look pixelated on HighDPI screens. We want to apply HighDPI
+    // manually in QOpenGLWidget.
+    // QCoreApplication::setAttribute(Qt::AA_EnableHighDpiScaling, false);
+    // #if QT_VERSION < QT_VERSION_CHECK(6, 0, 0)
+    //     QCoreApplication::setAttribute(Qt::AA_DisableHighDpiScaling);
+    // #endif
 
     vector<string> qtArgsStore;
     vector<char const*> qtArgsPtr;