+++ src/contour/opengl/TerminalWidget.h
@@ -25,6 +25,7 @@
 
 #include <terminal_renderer/Renderer.h>
 
+#include <QtCore/QFileSystemWatcher>
 #include <QtCore/QPoint>
 #include <QtCore/QTimer>
 #include <QtGui/QOpenGLExtraFunctions>
@@ -111,7 +112,8 @@ class TerminalWidget: public QOpenGLWidget, public TerminalDisplay, private QOpe
     void setWindowFullScreen() override;
     void setWindowMaximized() override;
     void setWindowNormal() override;
-    void setBackgroundBlur(bool _enable) override;
+    void setBlurBehind(bool _enable) override;
+    void setBackgroundImage(std::optional<terminal::BackgroundImage> const& backgroundImage) override;
     void toggleFullScreen() override;
     void setHyperlinkDecoration(terminal::renderer::Decorator _normal,
                                 terminal::renderer::Decorator _hover) override;
@@ -128,6 +130,10 @@ class TerminalWidget: public QOpenGLWidget, public TerminalDisplay, private QOpe
   public Q_SLOTS:
     void onFrameSwapped();
     void onScrollBarValueChanged(int _value);
+    void onRefreshRateChanged();
+    void onScreenDpiChanged();
+    void onScreenChanged();
+    void onDpiConfigChanged();
 
   signals:
     void terminalBufferChanged(terminal::ScreenType);
@@ -140,6 +146,9 @@ class TerminalWidget: public QOpenGLWidget, public TerminalDisplay, private QOpe
     //
     config::TerminalProfile const& profile() const noexcept { return session_.profile(); }
     terminal::Terminal& terminal() noexcept { return session_.terminal(); }
+    void configureScreenHooks();
+    void logDisplayInfo();
+    void watchKdeDpiSetting();
     terminal::PageSize screenSize() const
     {
         return screenSizeForPixels(pixelSize(), renderer_.gridMetrics());
@@ -168,7 +177,7 @@ class TerminalWidget: public QOpenGLWidget, public TerminalDisplay, private QOpe
     //
     TerminalSession& session_;
     std::function<void()> adaptSize_;
-    std::function<void(bool)> enableBackgroundBlur_;
+    std::function<void(bool)> enableBlurBehind_;
     terminal::renderer::Renderer renderer_;
     std::atomic<bool> initialized_ = false;
     bool renderingPressure_ = false;
@@ -180,6 +189,9 @@ class TerminalWidget: public QOpenGLWidget, public TerminalDisplay, private QOpe
     QTimer updateTimer_;
 
     RenderStateManager state_;
+
+    QFileSystemWatcher filesystemWatcher_;
+    crispy::Point lastScreenDPI_;
 
     // ======================================================================